Add support for ModUpdater
ModUpdater is a Fabric mod which automatically checks for and downloads
updates to other cooperating Fabric mods.  Adding this section of the
fabric.mod.json file tells it that this mod can be updated from
CurseForge and that it does not need to do strict version checking (so
any version of this mod for eg. 1.16 will work on 1.16.1 and 1.16.2 as
well).