Remove redundant stack trace printing for icon errors
I'm often in the situation where I'm developing a new mod that doesn't have an icon yet. I use Mod Menu to access my config file, and it prints a large stack trace twice. Once when entering, and once when exiting.

Since there is already a complete error line in red, which is acceptable to be shown I think, there is no need for a multi-line stack trace after that.

I've removed it for the square icon error as well, since that feels redundant too. But I can re-add that, let me know.

Thanks for the great mod!