Fix mod list selection highlight rendering going out of bounds
Fixes the scissoring not being applied to the selection highlight rendering by using the correct render pipeline which accepts scissor. I also decided to rewrite the code using DrawContext#fill as this is most likely easier to maintain (as Mojang will most likely rewrite the rendering again in the upcoming updates).