Add Feature: Mail-To Contacts (#868)
This commit adds support for contacting authors/contributors by email if they have added their contact information in accordance with the fabric.mod.json specification.
If an entry has contact information and if that contact information contains an email address, an email emoticon will be displayed next to their name in the menu.  Clicking a line with an email emoticon will prompt to open a mailto:address link.

- Add links to contacts' e-mail if present in fmj